backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
joblib==1.3.2
pydantic==2.5.0
python-dotenv==1.0.0
scipy==1.11.4
hsic==0.0.4
pyHSICLasso==0.0.6